requirements.txt
langchain
google-cloud-bigquery
google-cloud-aiplatform
streamlit
